Fix incompatible pointer type warning

+++ Limelight/Network/PairManager.m
@@ -202,7 +202,7 @@
     if (![resp isStatusOk]) {
         return false;
     } else {
-        long pairedStatus;
+        NSInteger pairedStatus;
         
         if (![resp getIntTag:@"paired" value:&pairedStatus]) {
             return false;
